SOC-2637 | discussion-reported-filter component
